Improve handling of missing error overlay
Sometimes `server/main.js` misses the elm error event on startup.
Previously, this would be detected and trigger a full page reload.
However, in some cases it could re-occur on the reload resulting in
five or six reloads before the error was displayed. This change
has the client send a message to the server to re-request the error
data instead of doing a full refresh.