Simplify client-server startup coordination
The browser client was configured to send a message to the server
to ask for error info if it detected that it missed them on the
initial websocket connection. This is now simplified so that the
client sends a message when it's ready. The server defers
sending anything until the client is ready, so there is no need
to check for missed messages.