Fix a bug in layouts with props using type variables
If the project has a layout with props using type variables, there's
an error in the code generation that throws:

```
Props type needs 1 argument, but I see 0 instead: in .elm-land/src/Layouts.elm
```

This commit updates the array of layout paths sent to the code
generator, including a boolean variable indicating if the layout uses Props
with type variables. In such case, the code generator creates valid
code for that layout, avoiding the current error.